Move general_hdffile to core and raname as _hdffile
