+++ src/ophyd_async/core/_hdffile.py
