+++ src/ophyd_async/epics/adcore/_hdf_writer.py
@@ -8,8 +8,8 @@ from ophyd_async.core import (DEFAULT_TIMEOUT, AsyncStatus, DetectorWriter,
                               NameProvider, PathProvider, ShapeProvider,
                               observe_value, set_and_wait_for_value,
                               wait_for_value)
+from ophyd_async.core._hdffile import _HDFDataset, _HDFFile
 
-from ._general_hdffile import _HDFDataset, _HDFFile
 from ._nd_file_hdf import NDFileHDF
 from ._nd_plugin import convert_ad_dtype_to_np
 from ._utils import FileWriteMode
