+++ tests/fastcs/panda/test_writer.py
@@ -7,7 +7,7 @@ import pytest
 from ophyd_async.core import (DEFAULT_TIMEOUT, Device, DeviceCollector,
                               SignalR, StaticFilenameProvider,
                               StaticPathProvider, set_mock_value)
-from ophyd_async.epics.adcore._general_hdffile import _HDFFile
+from ophyd_async.core._hdffile import _HDFFile
 from ophyd_async.epics.pvi import (create_children_from_annotations,
                                    fill_pvi_entries)
 from ophyd_async.fastcs.panda import (CommonPandaBlocks, DatasetTable,
